test new split order management
