+++ addons/point_of_sale/models/__init__.py
@@ -51,6 +51,5 @@ from . import product_tag
 from . import resource_calendar_attendance
 from . import product_uom
 from . import ir_sequence
-from . import pos_prep_order_group
 from . import pos_prep_order
 from . import pos_prep_line
