+++ addons/point_of_sale/models/pos_session.py
@@ -142,7 +142,7 @@ class PosSession(models.Model):
             'product.template.attribute.line', 'product.template.attribute.value', 'product.combo', 'product.combo.item', 'res.users', 'res.partner', 'product.uom',
             'decimal.precision', 'uom.uom', 'res.country', 'res.country.state', 'res.lang', 'product.category', 'product.pricelist', 'product.pricelist.item',
             'account.cash.rounding', 'account.fiscal.position', 'stock.picking.type', 'res.currency', 'pos.note', 'product.tag', 'ir.module.module', 'account.move',
-            'pos.prep.order.group', 'pos.prep.order', 'pos.prep.line']
+            'pos.prep.order', 'pos.prep.line']
 
     @api.model
     def _load_pos_data_domain(self, data, config):
