+++ addons/point_of_sale/security/ir.model.access.csv
@@ -57,4 +57,3 @@ access_pos_printer_pending_receipt,pos.printer.pending.receipt.user,model_pos_pr
 access_pos_printer_pending_receipt_manager,pos.printer.pending.receipt.manager,model_pos_printer_pending_receipt,point_of_sale.group_pos_manager,1,1,1,1
 access_preparation_line,pos.prep.line,model_pos_prep_line,point_of_sale.group_pos_user,1,1,1,1
 access_preparation_order,pos.prep.order,model_pos_prep_order,point_of_sale.group_pos_user,1,1,1,1
-access_preparation_order_group,pos.prep.order.group,model_pos_prep_order_group,point_of_sale.group_pos_user,1,1,1,1
