+++ addons/point_of_sale/static/src/app/models/data_service_options.js
@@ -9,14 +9,14 @@ export class DataServiceOptions {
                     record.finalized &&
                     typeof record.id === "number" &&
                     record.pos_session_id !== parseInt(odoo.pos_session_id) &&
-                    !record.prep_order_group_id,
+                    !record.prep_order_ids.length,
             },
             "pos.order.line": {
                 key: "uuid",
                 condition: (record) =>
                     record.order_id?.finalized &&
                     typeof record.order_id.id === "number" &&
-                    !record.order_id.prep_order_group_id,
+                    !record.order_id.prep_order_ids.length,
             },
             "pos.payment": {
                 key: "uuid",
@@ -30,18 +30,11 @@ export class DataServiceOptions {
             },
             "pos.prep.order": {
                 key: "uuid",
-                condition: (record) => !record.prep_order_group_id,
+                condition: (record) => !record.pos_order_id,
             },
             "pos.prep.line": {
                 key: "uuid",
                 condition: (record) => !record.prep_order_id,
-            },
-            "pos.prep.order.group": {
-                key: "uuid",
-                condition: (record) =>
-                    record.pos_order_ids.every(
-                        (order) => order.finalized && typeof order.id === "number"
-                    ),
             },
             "pos.pack.operation.lot": {
                 key: "uuid",
@@ -59,7 +52,6 @@ export class DataServiceOptions {
             "product.attribute.custom.value",
             "pos.prep.order",
             "pos.prep.line",
-            "pos.prep.order.group",
         ];
     }
 
@@ -70,7 +62,6 @@ export class DataServiceOptions {
             "pos.order.line": ["uuid"],
             "pos.payment": ["uuid"],
             "product.attribute.custom.value": ["uuid"],
-            "pos.prep.order.group": ["uuid"],
             "pos.pack.operation.lot": ["uuid"],
             "product.template": ["pos_categ_ids", "write_date"],
             "product.product": ["pos_categ_ids", "barcode"],
