+++ addons/point_of_sale/static/tests/unit/data/generate_model_definitions.js
@@ -51,7 +51,6 @@ import { BarcodeNomenclature } from "./barcode_nomenclature.data";
 import { ProductAttributeValue } from "./product_attribute_value.data";
 import { PosPrepOrder } from "./pos_prep_order.data";
 import { PosPrepLine } from "./pos_prep_line.data";
-import { PosPrepOrderGroup } from "./pos_prep_order_group.data";
 
 export const hootPosModels = [
     ResCountry,
@@ -104,7 +103,6 @@ export const hootPosModels = [
     ResourceCalendar,
     BarcodeNomenclature,
     PosPrepOrder,
-    PosPrepOrderGroup,
     PosPrepLine,
 ];
