--- addons/point_of_sale/static/tests/unit/data/pos_prep_order_group.data.js
@@ -1,9 +1,0 @@
-import { models } from "@web/../tests/web_test_helpers";
-
-export class PosPrepOrderGroup extends models.ServerModel {
-    _name = "pos.prep.order.group";
-
-    _load_pos_data_fields() {
-        return [];
-    }
-}
