+++ addons/pos_restaurant/static/src/app/screens/product_screen/actionpad_widget/actionpad_widget.xml
@@ -50,7 +50,7 @@
                 <button
                     class="button w-50 btn btn-secondary btn-lg d-flex flex-row align-items-center justify-content-center"
                     t-on-click="() => doSubmitOrder.call()"
-                    t-elif="!ui.isSmall and this.pos.unwatched.printers.length"
+                    t-elif="!ui.isSmall and this.pos.unwatched.printers.length  and this.currentOrder.uiState.printStack.length"
                     t-att-disabled="doSubmitOrder.status === 'loading'">
                     Reprint
                 </button>
