+++ addons/pos_restaurant/static/src/app/screens/tip_screen/tip_screen.js
@@ -94,6 +94,7 @@ export class TipScreen extends Component {
         const serializedTipLine = order.getSelectedOrderline().serializeForORM();
         order.getSelectedOrderline().delete();
         delete serializedTipLine["record_dependencies"];
+        delete serializedTipLine["relations_uuid_mapping"];
         const serverTipLine = await this.pos.data.create("pos.order.line", [serializedTipLine]);
         await this.pos.data.write("pos.order", [serverId], {
             is_tipped: true,
