+++ addons/pos_restaurant/static/tests/tours/utils/product_screen_util.js
@@ -30,12 +30,12 @@ export function guestNumberIs(num) {
         },
     ];
 }
-export function OrderButtonNotContain(data) {
+export function OrderButtonContain(data) {
     const steps = [
         {
             isActive: ["desktop"],
             content: "check order button not contain data",
-            trigger: `.product-screen .submit-order:not(:contains("${data}"))`,
+            trigger: `.product-screen .submit-order:contains("${data}")`,
             run: function () {}, // it's a check
         },
     ];
