+++ addons/pos_self_order/controllers/orders.py
@@ -70,9 +70,8 @@ class PosSelfOrderController(http.Controller):
             'pos.payment': self.env['pos.payment']._load_pos_self_data_read(order.payment_ids, config),
             'pos.payment.method': self.env['pos.payment.method']._load_pos_self_data_read(order.payment_ids.mapped('payment_method_id'), config),
             'product.attribute.custom.value': self.env['product.attribute.custom.value']._load_pos_self_data_read(order.lines.custom_attribute_value_ids, config),
-            'pos.prep.order': self.env['pos.prep.order']._load_pos_data_read(order.prep_order_group_id.prep_order_ids, config) if config else [],
-            'pos.prep.line': self.env['pos.prep.line']._load_pos_data_read(order.prep_order_group_id.prep_order_ids.prep_line_ids, config) if config else [],
-            'pos.prep.order.group': self.env['pos.prep.order.group']._load_pos_data_read(order.prep_order_group_id, config) if config else [],
+            'pos.prep.order': self.env['pos.prep.order']._load_pos_data_read(order.prep_order_ids, config) if config else [],
+            'pos.prep.line': self.env['pos.prep.line']._load_pos_data_read(order.prep_order_ids.prep_line_ids, config) if config else [],
         }
 
     def _verify_line_price(self, lines, pos_config, preset_id):
