+++ addons/pos_self_order/models/__init__.py
@@ -18,4 +18,3 @@ from . import res_country
 from . import mail_template
 from . import pos_prep_order
 from . import pos_prep_line
-from . import pos_prep_order_group
