+++ addons/pos_self_order/static/tests/unit/data/pos_session.data.js
@@ -46,7 +46,6 @@ patch(PosSession.prototype, {
             "res.country",
             "res.country.state",
             "pos.prep.order",
-            "pos.prep.order.group",
             "pos.prep.line",
         ];
     },
